Remove section numbers from export
Use org-mode export directive `num:nil' to remove section numbers
during export. Not sure whether GitHub recognizes these export
directives but if it does this would likely entail its using `ox-gfm'
or the package is based of the gh implementation.
That may or may not be significant but nonetheless would be interesting...