Replace fork with emacsmirror repository in org recipe
The emacsmirror repo removes the need to maintain my own fork. The
only modification I made to org-mode was a patch to fontify export
latex src blocks and this feature is already provided by the
magic-latex-buffer package. My fork also takes forever to clone and
the recipe from emacsmirror does not. This is the same recipe used by
straight.el, but the :files and :local-repo-name keywords are still
necessary to use packages in the org contrib/ directory, and it
appears that straight.el replaces the default recipe with a custom one
rather than merge so the :host and :repo keywords for straight.el to
recognize the recipe as a git backend.

One item of note: Easy template expansion breaks with the emacsmirror
clone. The docstring for `org-structure-template-alist' mentions
that the variable was changed or added in v9.2. This is probably the
reason for the breaking change. The url mentioning this in the
changelog is broken but the release notes also mention that the command
`org-insert-structure-template' was added in v9.2, so it's possible that
templating is now done by insertion rather than text expansion.