Fix indentation levels of src blocks, Add counsel-dash replacement
The counsel-dash package has a lot of overhead because it's just a
wrappper around the Helm equivalent. Also I can't get it to work
outside of Spacemacs. It's fairly simple to implement the
`dash-at-point' functionality and use the dash:// protocol if you have
a lisence to open the docset in the external application.