Refactor for compatibility with straight.el package management system
Previously I was using the straight.el package mantainer's fork of
use-package.
This worked for a while but isn't very suitable for long-term
stability.
The variable `straight-use-package-by-default' was added to default
to the `straight.el' integration of the `use-package' library, but
I'm having compilation issues with that. Replacing all the `ensure'
keywords in the `use-package' blocks with the `straight' keyword
essentially opts-in manually to the straight.el version of
`use-package'.