Ensure that gpg is installed before checking for keys
If gpg is not installed then a string will still be printed. Checking
for a string is definitely not the best way to do this because it
still does not account for possible edge cases where the executable is
found, the keys are not, and a string (e.g an error message) is still
printed but this works for the time being. Also, fix error in the
flag being passed to the gpg program in the inferior shell command.