Fix per-package capture templates
Appending to capture templates list in lazily-loaded package
configuration blocks does not work. Demanding the package instead
seems to fix this problem. I'm worried that this will trigger the
built-in org to be loaded instead of the emacsmirror so I'm also using
the :after key even though the entire org-mode configuration is
already wrapped in a with-eval-after-load sexp in the section noweb block