Revert built-in pseudo packages supplied by straight.
Built-in packages configured by way of use-package declarations need
to be here for compatibility with the develop branch of
straight.el. Another option is to set the `:straight' key for these
packages to nil as I was doing previously. I modified this strategy in
hopes that it would fix the lag issue with the popup buffer that
occurs when instructing straight to read version lockfiles and revert
packages. I thought I had this working in commit ffe71a8.

As described in the previous commit, the buffer lag is still a
problem, but I think the elision of the straight key for packages is
more intuitive, if not explicit.

This reverts commit 9edd4a1479b405b41a9b8585d2f61b5b18f21746.