Use pdf-view-mode to view TeX documents
The default viewer specifications are determined based on the
current operating system, in `TeX-view-program-list-builtin'
variable. AucTeX provides an implementation for using pdf-tools, but
this is not enabled on macOS or Windows, for whatever reason. I'm
adding it to the user-facing equivalent, `Tex-view-program-list', and
overriding the default by setting it as the value of the pdf
predicate in the `TeX-view-program-selection' alist. The overall
effect of these modifications is that the pdf outputted as a result of
executing the `Tex-view' (C-c C-v) command in LaTeX major mode will
now be opened in a `pdf-view-mode' buffer rather than the system viewer
by default, regardless of operating system.

Note that the described behavior has to be implemented as a LaTeX-mode
hook to have any effect. I'm not entirely sure why this is, but AucTeX
has a fairly complex loading process so that's a curiosity for another
day.

Adding  `TeX-revert-document-buffer' to the
`TeX-after-compilation-finished-functions' hook automates the viewer
updating upon re-compilation. This is the recommendation when using
emacs-internal viewers in the latter variable docstring.

----

I also modified the theme implementation to enable my
atom-one-dark customizations after that theme is loaded. This process
still has to be done twice in order for `org-beautify-theme' to play
nicely - once at startup and then again after initialization - for
reasons that continue to mystify me.

---

I think there may be a complication with ggtags-mode due to my current
emacs being compiled with ctags enabled. Might need to recompile with
the following option, from the spacemacs LaTeX layer documentation:

./configure --program-transform-name='s/^ctags$/ctags.emacs/'

---

Moved pdf-view settings to from org-pdfview configuration to the
pdf-tools config. Either way works because the former won't load until
pdf-view-mode bindings are defined, but pdf-tools provides this library.